Check for contact_id before attempting to find account card
No need to call the API again if there the method is unable to find a contact_id based on the param hash it receives.

Method now throws an error if it receives a param hash with no valid contact id OR shopper id OR card id, which means it can be used by Validator methods.