Add dry run logic
Simulate code/card/contact transactions by calling the '/dryRun' endpoint. This can be used to display a charge summary to the end customer for the purpose of asking them for confirmation to proceed with the transaction.

'NSF' is set to 'false' by default: in the case that a card has insufficient funds for a (simulated) transaction, instructs the API to return a simulated transaction for the maximum available amount on the card rather than responding with an error. Necessary for accurate context-specific balance-checking.

Note that for the Contact method, as with other Contact methods, this one is a wrapper for the Card method.