Update JWT generation (breaking change)
Updates token generator to take any contact identifier (shopperId, contactId, or contact userSuppliedId) and generate a properly signed shopper token.
Breaks with previous implementation because the contact identifier is now expected to be a value in a hash, with the key indicating which identifier it is (ie: {shopper_id: 'cust-123'}). }
Also now adds 'iss: MERCHANT' to the payload.

These changes bring it in line with updated API requirements and with the other client libraries.